feat: improve typings
