+++ src/components/dashboard.tsx
@@ -23,152 +23,7 @@ import { categoryIcons, statusIcons } from '@/types/types';
 export default function DashboardPage({ sessionId, user, allProjects, allDonations }: { sessionId: string; user: TUser | TOrganization; allProjects: TProject[]; allDonations: any[]; }) {
   console.warn(sessionId);
 
-  const [projects, setProjects] = useState(allProjects);
-  const [donations] = useState(allDonations);
-  const [editingProject, setEditingProject] = useState<TProject | null>(null);
-  const [payoutProject, setPayoutProject] = useState<any>(null);
-
-  const handleEditProject = (project: any) => {
-    setEditingProject(project);
-  };
-
-  const handleUpdateProject = (updatedProject: any) => {
-    setProjects(projects.map(p => (p.id === updatedProject.id ? updatedProject : p)));
-    setEditingProject(null);
-  };
-
-  const handlePayoutRequest = (project: any) => {
-    setPayoutProject(project);
-  };
-
-  const handlePayoutComplete = (project: any) => {
-    // In a real application, you would update the project status or payout status here
-    console.log(`Payout requested for project: ${project.title}`);
-    setPayoutProject(null);
-  };
 
   return (
-    <div className="space-y-8">
-      <Card>
-        <CardHeader>
-          <CardTitle>My Projects</CardTitle>
-        </CardHeader>
-        <CardContent>
-          <Tabs defaultValue="active">
-            <TabsList>
-              <TabsTrigger value="active">Active</TabsTrigger>
-              <TabsTrigger value="completed">Completed</TabsTrigger>
-              <TabsTrigger value="expired">Expired</TabsTrigger>
-            </TabsList>
-            {['active', 'completed', 'expired'].map(status => (
-              <TabsContent key={status} value={status}>
-                {projects
-                  .filter(project => project.status === status)
-                  .map(project => (
-                    <Card key={project.id} className="mb-4">
-                      <CardContent className="p-4">
-                        <div className="mb-2 flex items-center justify-between">
-                          <h3 className="text-lg font-semibold">{project.title}</h3>
-                          <div className="flex items-center">
-                            {React.createElement(
-                              categoryIcons[project.category as string] || Users, // Provide a fallback icon
-                            )}
-                            <span className="text-sm">{project.category}</span>
-                          </div>
-                        </div>
-                        <p className="mb-4 text-sm text-muted-foreground">{project.description}</p>
-                        <div className="mb-2 flex items-center justify-between">
-                          <span className="text-sm font-medium">
-                            $
-                            {project.raised.toLocaleString()}
-                            {' '}
-                            raised of $
-                            {project.goal.toLocaleString()}
-                          </span>
-                          <span className="text-sm">
-                            {project.endDate > 0 ? `${project.daysLeft} days left` : 'Ended'}
-                          </span>
-                        </div>
-                        <Progress value={(Number(project.raised) / Number(project.goal)) * 100} className="mb-2" />
-                        <div className="flex items-center justify-between">
-                          <span className="text-sm text-muted-foreground">
-                            {React.createElement(
-                              statusIcons[project.status] || AlertCircle, // Provide a fallback icon
-
-                            )}
-                            {status.charAt(0).toUpperCase() + status.slice(1)}
-                          </span>
-                          <div>
-                            {status === 'active' && (
-                              <>
-                                <Button
-                                  size="sm"
-                                  variant="outline"
-                                  className="mr-2"
-                                  onClick={() => handleEditProject(project)}
-                                >
-                                  <Edit className="mr-1 size-4" />
-                                  Edit
-                                </Button>
-                                <Button size="sm">Manage Project</Button>
-                              </>
-                            )}
-                            {status === 'completed' && (
-                              <Button size="sm" onClick={() => handlePayoutRequest(project)}>
-                                <DollarSign className="mr-1 size-4" />
-                                Request Payout
-                              </Button>
-                            )}
-                            {status === 'expired' && (
-                              <Button size="sm" variant="outline" disabled>
-                                Expired
-                              </Button>
-                            )}
-                          </div>
-                        </div>
-                      </CardContent>
-                    </Card>
-                  ))}
-              </TabsContent>
-            ))}
-          </Tabs>
-          <Button className="mt-4">Create New Project</Button>
-        </CardContent>
-      </Card>
-
-      <Card>
-        <CardHeader>
-          <CardTitle>My Donations</CardTitle>
-        </CardHeader>
-        <CardContent>
-          <ul>
-            {donations.map(donation => (
-              <li key={donation.id} className="mb-2 rounded-md bg-secondary p-2">
-                <span className="font-semibold">{donation.projectTitle}</span>
-                <div className="flex justify-between text-sm text-muted-foreground">
-                  <span>
-                    $
-                    {donation.amount}
-                  </span>
-                  <span>{donation.date}</span>
-                </div>
-              </li>
-            ))}
-          </ul>
-        </CardContent>
-      </Card>
-
-      {editingProject && (
-        <EditProjectOverlay
-          project={editingProject}
-          onClose={() => setEditingProject(null)}
-          onUpdate={handleUpdateProject}
-        />
-      )}
-
-      {payoutProject && (
-        <PayoutOverlay project={payoutProject} onClose={() => setPayoutProject(null)} onPayout={handlePayoutComplete} />
-      )}
-    </div>
-  );
+    
 }
